+++ src/plugins/optgroup_columns/plugin.js
@@ -1,6 +1,7 @@
 /**
  * Plugin: "optgroup_columns" (selectize.js)
  * Copyright (c) 2013 Simon Hewitt & contributors
+ * Copyright (c) 2020 Selectize Team & contributors* 
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this
  * file except in compliance with the License. You may obtain a copy of the License at: