+++ src/plugins/remove_button/plugin.js
@@ -1,6 +1,7 @@
 /**
  * Plugin: "remove_button" (selectize.js)
  * Copyright (c) 2013 Brian Reavis & contributors
+ * Copyright (c) 2020 Selectize Team & contributors
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this
  * file except in compliance with the License. You may obtain a copy of the License at:
@@ -109,6 +110,7 @@ Selectize.define('remove_button', function(options) {
 						if (self.deleteSelection()) {
 							self.setCaret(self.items.length);
 						}
+						return false;
 					});
 
 				};