+++ src/utils.js
@@ -189,6 +189,10 @@ var watchChildEvent = function($parent, event, selector, fn) {
  */
 var getSelection = function(input) {
 	var result = {};
+  if(input === undefined) {
+    console.warn('WARN getSelection cannot locate input control');
+    return result;
+  }
 	if ('selectionStart' in input) {
 		result.start = input.selectionStart;
 		result.length = input.selectionEnd - result.start;
@@ -238,12 +242,17 @@ var measureString = function(str, $parent) {
 	if (!Selectize.$testInput) {
 		Selectize.$testInput = $('<span />').css({
 			position: 'absolute',
-			top: -99999,
-			left: -99999,
 			width: 'auto',
 			padding: 0,
 			whiteSpace: 'pre'
-		}).appendTo('body');
+		});
+
+		$('<div />').css({
+			position: 'absolute',
+			width: 0,
+			height: 0,
+			overflow: 'hidden'
+		}).append(Selectize.$testInput).appendTo('body');
 	}
 
 	Selectize.$testInput.text(str);